add typesafe-i18n safe svelte adapter
for a bit of context on this one: the official ts-in svelte adapter is not SSR-safe. it works as expected for fully synchronous code, but when eg a load function takes a bit and a new request comes in, the global i18n variables (yikes) get changed and both requests get the same i18n object. I wrote another adapter that uses the page store to avoid this leak.